refactor(card): apply theme
