+++ src/index.ts
@@ -6,7 +6,7 @@ export * from './components/NumberInput/NumberInput';
 export * from './components/Checkbox/Checkbox';
 export * from './components/Chip/Chip';
 export * from './components/Select/Select';
-export * from './components/Card/Card';
+export * from './components/Card';
 export * from './components/Radio/Radio';
 export * from './components/Tabs/Tabs';
 export * from './components/Table/Table';
